% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_examples.R
\name{ConvertExamplesToTextFiles}
\alias{ConvertExamplesToTextFiles}
\title{Convert the word files to plain txt files using the python 
function supplied with this package.}
\usage{
ConvertExamplesToTextFiles(path, txt_path)
}
\arguments{
\item{path}{the path to the folder including the examples}

\item{txt_path}{path to the folder where the txt files will be saved}
}
\description{
Convert the word files to plain txt files using the python 
function supplied with this package.
}
