% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_examples.R
\name{GetExamplesFromTextFiles}
\alias{GetExamplesFromTextFiles}
\title{Collect all examples from the text files and return them as a data frame (unprocessed)}
\usage{
GetExamplesFromTextFiles(path)
}
\arguments{
\item{path}{path to the folder where the txt files will be saved}
}
\value{
a data frame consisting of all the examples in the specified folder
}
\description{
Collect all examples from the text files and return them as a data frame (unprocessed)
}
