% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{GetPatternMatrix}
\alias{GetPatternMatrix}
\title{Get a matrix between patterns organized in a data frame
with only categorical variables.
Expects the column identifying the patterns to be the first one}
\usage{
GetPatternMatrix(mydf)
}
\arguments{
\item{mydf}{the dataframe}
}
\description{
Get a matrix between patterns organized in a data frame
with only categorical variables.
Expects the column identifying the patterns to be the first one
}
