% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{GetRandomSubset}
\alias{GetRandomSubset}
\title{Collect a subset for random selection}
\usage{
GetRandomSubset(mydf, parametry.1, parametry.2, n)
}
\arguments{
\item{parametry.1}{list (attr=x, values=x)}

\item{parametry.2}{list (attr=x, values=x)}
}
\description{
Collect a subset for random selection
}
