% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forests.R
\name{GetTree}
\alias{GetTree}
\title{Convenience function for getting a cond. inference tree}
\usage{
GetTree(seed = F)
}
\arguments{
\item{seed}{If you want to fix the results so that they will be consistent
between each rounds}
}
\description{
Convenience function for getting a cond. inference tree
}
