% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PrintProfilePlot}
\alias{PrintProfilePlot}
\title{Prints a plot comparing the profiles of different adjectives}
\usage{
PrintProfilePlot(abs_table, adjectives, min_pc = 4)
}
\arguments{
\item{abs_table}{cross-tabulated table with absolute numbers}

\item{adjectives}{a vector consisting of th adjectives / constructions you want to compare}

\item{min_pc}{minimum percentage value to be printed as text on the plot}
}
\description{
Prints a plot comparing the profiles of different adjectives
}
\examples{

# Example 1

library(stockholm)
bolshaja$otdelno  <- apply(bolshaja,1,function(r)gsub("\\\\[adj\\\\]",r[["adj"]],r[["pattern"]])) 
bolshaja_moj <- subset(bolshaja, grepl("moj",pattern))
bolshaja_bez_moj <- subset(bolshaja, !grepl("moj",pattern))
tablica_s_absoljutnymi_tsiframi_moj <- table(bolshaja_moj$participants, bolshaja_moj$otdelno)
tablica_s_absoljutnymi_tsiframi_bez_moj <- table(bolshaja_bez_moj$participants, bolshaja_bez_moj$otdelno)
PrintProfilePlot(tablica_s_absoljutnymi_tsiframi_moj,c("milyj moj","moj milyj","dorogoj moj","moj dorogoj"))


}
