% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_examples.R
\name{ProcessExamples}
\alias{ProcessExamples}
\title{Process the collected examples, add variables etc}
\usage{
ProcessExamples(examples)
}
\arguments{
\item{examples}{a data frame produced by GetExamplesFromTextFiles}
}
\value{
a data frame consisting of all the examples in the specified folder
}
\description{
Here we try to separate the participiants of the speech situation, the
pattern (whether there is a pronoun etc ) based on the information
originally present in the text file
}
