% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{sozdai_barplot_s_prilagatelnym}
\alias{sozdai_barplot_s_prilagatelnym}
\title{Returns a barplot of adjectives}
\usage{
sozdai_barplot_s_prilagatelnym(
  dannyje,
  this_adj,
  spisok_utshastnikov = c(),
  protsenty = F,
  freqlabels = F,
  nlabels = F,
  y_axis_all_percents = F
)
}
\arguments{
\item{dannyje}{bolshaja, bolshaja_moj, bolshaja_bez_moj etc..}

\item{spisok_utshastnikov}{vector of participants to include in the plot (if not set, then all)}

\item{protsenty}{whether to use percentages instead of counts}

\item{freqlabels}{whether to print labels of percentages on the plot}

\item{nlabels}{whether to print labels of counts on the plot}

\item{y_axis_all_percents}{whether to use scale of 0 to 100 with percentages}

\item{adj}{the adjective ('dorogoj', 'milyj'...)}
}
\description{
Returns a barplot of adjectives
}
\examples{


# samyj prostoi:
sozdai_barplot_s_prilagatelnym(bolshaja, 'dorogoj')

# esli hotshetsa ogranitsit spisok utshastnikoc
sozdai_barplot_s_prilagatelnym(bolshaja, 'dorogoj', spisok_utshastnikov=c('niania', 'snizu vverh', 'zhivotnye'))

# esli hotshetsa pokazat protsenty
sozdai_barplot_s_prilagatelnym(bolshaja, 'dorogoj', protsenty=T)

# ili mozhet byt lutshe vot tak
sozdai_barplot_s_prilagatelnym(bolshaja, 'dorogoj', protsenty=T, y_axis_all_percents=T)

# potom mozhno eshe napisat tsifry nad barami
sozdai_barplot_s_prilagatelnym(bolshaja, 'dorogoj', protsenty=T, y_axis_all_percents=T, freqlabels=T)

# ...ili eshe
sozdai_barplot_s_prilagatelnym(bolshaja, 'dorogoj', protsenty=T, y_axis_all_percents=T, freqlabels=T, nlabels=T)

}
