% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{sozdai_tablitsa_s_prilagatelnym}
\alias{sozdai_tablitsa_s_prilagatelnym}
\title{Returns a table filtered by adj}
\usage{
sozdai_tablitsa_s_prilagatelnym(dannyje, this_adj)
}
\arguments{
\item{dannyje}{bolshaja, bolshaja_moj, bolshaja_bez_moj etc..}

\item{this_adj}{the adjective in question (dorogoj, milyj...)}
}
\description{
Returns a table filtered by adj
}
\examples{

# Example 1:

tablitsa  <-  sozdai_tablitsa_s_prilagatelnym(bolshaja, 'dorogoj')


}
