% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{sozdai_tablitsa_s_utshastnikamki}
\alias{sozdai_tablitsa_s_utshastnikamki}
\title{Returns a table filtered by participants}
\usage{
sozdai_tablitsa_s_utshastnikamki(dannyje, spisok_utshastnikov)
}
\arguments{
\item{dannyje}{bolshaja, bolshaja_moj, bolshaja_bez_moj etc..}

\item{spisok_utshastnikov}{list of participants to include in the table}
}
\description{
Returns a table filtered by participants
}
\examples{

# Example 1:

tablitsa  <-  sozdai_tablitsa_s_utshastnikamki(bolshaja_bez_moj, c('roditeli deti', 'deti roditeli'))
PrintProfilePlot(tablitsa ,c("dorogoj", "milyj", "rodnoj", "lubimyj"), vertical=T) 


}
