% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{GetSilhouettes}
\alias{GetSilhouettes}
\title{Compute the widths of the silhouttes of different clustering solutions}
\usage{
GetSilhouettes(myhc, mydist, maxclust)
}
\arguments{
\item{myhc}{a hierarchically clustered object}

\item{mydist}{a distance matrix on which the clustering was based}

\item{maxclust}{maximum number of clusters}
}
\description{
Compute the widths of the silhouttes of different clustering solutions
}
