% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{GetSilhouettes}
\alias{GetSilhouettes}
\title{Compute the widths of the silhouttes of different clustering solutions}
\usage{
GetSilhouettes(myhc, mydist, maxclust)
}
\arguments{
\item{myhc}{a hierarchically clustered object}

\item{mydist}{a distance matrix on which the clustering was based}

\item{maxclust}{maximum number of clusters}
}
\description{
Compute the widths of the silhouttes of different clustering solutions
}
\examples{

# Example 1
bolshaja <- readRDS("tablica_?_?_2018.rds")
bolshaja.subset  <- bolshaja[,c("adj","participants","pattern","tonalnost")]
bolshaja.matrix <- GetPatternMatrix(bolshaja.subset)
bolshaja.dist <- dist(bolshaja.matrix, method="canberra")
bolshaja.hc <- hclust(bolshaja.dist,method="ward.D2")
silhouette.widths <- GetSilhouettes(bolshaja.hc, bolshaja.dist,length(unique(bolshaja.subset$adj))-1)
plot(bolshaja.hc)

# Example 2
bolshaja <- readRDS("tablica_?_?_2018.rds")
bolshaja.subset  <- bolshaja[,c("adj","participants","pattern","tonalnost")]
bolshaja.subset$adj <- as.character(bolshaja.subset$adj)
bolshaja.subset <- subset(bolshaja.subset,adj != "ljubimyj")
bolshaja.subset$adj <- apply(bolshaja.subset,1,function(r)gsub("\\\\[adj\\\\]",r["adj"],r["pattern"],))
bolshaja.subset  <- bolshaja.subset[,c("adj","participants")]
bolshaja.matrix<-GetPatternMatrix(bolshaja.subset)
bolshaja.dist <- dist(bolshaja.matrix, method="canberra")
bolshaja.hc <- hclust(bolshaja.dist,method="ward.D2")
silhouette.widths <- GetSilhouettes(bolshaja.hc, bolshaja.dist,length(unique(bolshaja.subset$adj))-1)
plot(bolshaja.hc)


# Example 3
bolshaja <- readRDS("tablica_?_?_2018.rds")
bolshaja.subset  <- bolshaja[,c("adj","participants","pattern","tonalnost")]
bolshaja.subset$adj <- apply(bolshaja.subset,1,function(r)gsub("\\\\[adj\\\\]",r["adj"],r["pattern"],))
bolshaja.subset  <- bolshaja.subset[,c("adj","participants")]
bolshaja.matrix<-GetPatternMatrix(bolshaja.subset)
bolshaja.dist <- dist(bolshaja.matrix, method="canberra")
bolshaja.hc <- hclust(bolshaja.dist,method="ward.D2")
silhouette.widths <- GetSilhouettes(bolshaja.hc, bolshaja.dist,length(unique(bolshaja.subset$adj))-1)
plot(bolshaja.hc)


# Example 4
bolshaja <- readRDS("tablica_?_?_2018.rds")
bolshaja.subset  <- bolshaja[,c("adj","participants","pattern","tonalnost")]
bolshaja.subset$adj <- as.character(bolshaja.subset$adj)
bolshaja.subset <- subset(bolshaja.subset,adj != "ljubimyj")
bolshaja.subset$adj <- apply(bolshaja.subset,1,function(r)gsub("\\\\[adj\\\\]",r["adj"],r["pattern"],))
bolshaja.matrix<-GetPatternMatrix(bolshaja.subset)
bolshaja.dist <- dist(bolshaja.matrix, method="canberra")
bolshaja.hc <- hclust(bolshaja.dist,method="ward.D2")
silhouette.widths <- GetSilhouettes(bolshaja.hc, bolshaja.dist,length(unique(bolshaja.subset$adj))-1)
plot(bolshaja.hc)

}
